Decide on reconnection scheme
Reconnection will be using a package-level .DB_ENV environment that contains a connection pool object that represents a single backend connection to the DB. To differentiate between possible different backends, a hash based on the database filepath is used as an abbreviated identifier.

At the start of most queries, objects will check if their connections to the backend are valid. If not, they will pull from the hash associated with them in .DB_ENV